Add tests for set_score and fix weight.
Gradescope's decorators add attributes directly to the test methods.
Unfortunately, this approach doesn't work when the tests are also
wrapped by the parameterized module.  Constant values like weight can
work as long as all tests have the same weight, but values like score,
which is used for partial credit, need to be determined at run time.
This is a problem because parameterized does not provide a way for the
wrapped function to access its wrapper, so th wrapper's attributes
cannot be changed.

This commit solves the problem by adding a __gradescope__ dictionary
directly to the test's class to track gradescope variables.  A class
cleanup function is then injected into the class that maps those
variables to the appropriate test after all the tests have run.  The
mapping assumes all the tests in the class run, and that they run in the
default order.  If this assumption causes problems, it may be possible
to examine the test suites's results to determine fix mapping issues.