Use dot notation for modules in Options.
Importer expects a module, not a path.  This means we will need to convert it to a path for tests that don't use Importer, but if we took a path instead, we would have to convert it to a module for Importer.  There's a conversion required either way.  The attributes were already named ref_module and sub_module, so I chose default as module notation, and convert to path where needed.